backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.4
pydantic-settings==2.7.1

# Firebase / Firestore
firebase-admin==6.6.0
google-cloud-firestore==2.20.1

# Vertex AI — replaces google-generativeai (consumer SDK)
google-cloud-aiplatform>=1.60.0

# GCP services
google-cloud-secret-manager>=2.20.0
google-cloud-storage>=2.14.0
google-cloud-bigquery>=3.14.0
google-cloud-tasks>=2.16.0
google-cloud-logging>=3.9.0

# Web Scraping
httpx==0.28.1
beautifulsoup4==4.12.3
lxml==5.3.0

# CSV/Data
pandas==2.2.3
numpy>=1.26.0

# Auth
PyJWT==2.10.1
bcrypt==5.0.0

# Observability
opentelemetry-sdk>=1.22.0
opentelemetry-exporter-gcp-trace>=1.6.0
opentelemetry-instrumentation-fastapi>=0.43b0

# Retry / resilience
tenacity>=8.2.0

# Redis cache
redis>=5.0.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
